UI: All the rest of the stuff
DEVISE
UI for:
- Registrations
Removed the Cancel my Account feature

CLAIMS
UI for:
- all claims html

DEPOSIT
UI for:
- all html
- added require for application.js so that editdeposit can use wizard.js
... cause the form looks so awkwardly empty, that I decided to give it a
step by step with wizard

ACCOUNTS
UI for:
- viewaccount
container for Justification

BUDGET
UI for:
- main
datatables can pagination, filter and sort, but the horizontal scrolling
not so good. If overflow, no horizontal scroll. There issome javascript
kind of remedy but Im not too sure of it.
So, I combined 4 columns into one, called Options: It is a dropdown menu
incorporated into the table header that displays the meaning of the 4
icons in the cell below it.
- reviewmain
Same for this.
- show
same
For the Budget html with the 13 categories, I placed the table of 13 cat
at the bottom.