More UI
In Fullsummary, I moved the allocate and pullback buttons into the
container on top. There is a checkbox option to expand the container.

In the rest of the budget files, changed forms to modals. Jinq, I need
you to test on MacBook Safari and Google Chrome, Safari seems to be
anti-social from last UAT

Pullback and setcontributionrate has UI now.

The flash[:error] in viewclaim and viewdeposit are nowin a modal. You
can try. Same for the alert and notice in application.html, meaning when
you sign in, instead of a red line saying successful, a modal pops up