UI for new recommendation
I have implemented a token counter, such that when you make a choice,
the token will decrease, starting from ten. it uses javascript too. So
its together with the validate script. I also made the document call
that counter whenever it loads, such that even by clicking back, the
counter will calculate again.

I am not sure if the same can be done for endorsing forms. forcing the
page to reload on click.